Abstract Feed Handling
Adds an abstract feed model to serve as the basis of other feed models.
The abstract model will be responsible for applying the callbacks to a
collection of items and using the data returned from the callbacks to
build the CSV file. It expects the concrete classes to implement methods
for getting the callbacks for that feed and the items to be included in
the feed.