Move Callback Mapping Config
Move the callback mappings from feed specific to a shared mapping for
all feeds. Replace the feed specific mappings with a simple comma
separated list of fields to include in that feed. The abstract model
will join the fields to be included in that feed with the combined list
of callbacks to get the callback to apply to that field.

Add config helper methods for getting the combined callback mapping.