Fix Abstract Feed
Maps feed fields to callback mappings when getting headers and apply
same validation on callback configurations. This ensures that any fields
data is extracted for are also present or not present in the column
headers.

Require all callback configurations to include a 'column_name'. Add this
check to the validation method.

Fix some related tests.